[Plataforma Python] Se incorpora normalizacion y cuantizacion de las entradas en punto fijo. La nueva funcion es NormalizedFixed_Matrix.py que se incorporo en la Plataforma_Python_Punto_Flotante.ipynb para dar lugar a una nueva plataforma de punto fijo Plataforma_Python_Punto_Fijo.ipynb . Pendiente de incorporar cuantizacion en la funcion CounterGenSysMult_V2.py y desnormalizacion. Por otro lado falta loguear las entradas y salidas para testbench y vectormatching respectivamente.

+++ Python_HW_Verification/InputGenerator.py
@@ -14,8 +14,8 @@ def InputGenerator(premult_A,postmult_B):
     N_matrix_mult = len(postmult_B)//len(premult_A)
     print('N_matrix_mult: ',N_matrix_mult)
     
-    premult_A_stream   = np.zeros((N, (N_matrix_mult*M + M -1)), dtype=int)
-    postmult_B_stream  = np.zeros(((N_matrix_mult*N + N -1, M)), dtype=int)
+    premult_A_stream   = np.zeros((N, (N_matrix_mult*M + M -1)), dtype=float)
+    postmult_B_stream  = np.zeros(((N_matrix_mult*N + N -1, M)), dtype=float)
     
     for i in range(N):
         a = premult_A[i,:]                                  # Obtengo la primer fila
